Avoid flashing the wrong color theme on load
The color theme is set via the "dark" class on the html root element.
This is handled in JS code. However,
since the JS is loaded after the html content, there is a short delay
before the theme is applied as expected.

This commit fixes this issue by adding blocking JS that checks for dark
mode before the page's  first render.