Reduce layout jump when rendering email link
Since the email link is not rendered to the static html, the renderer
does not allocate space for it during the initial load. This causes a
noticeable layout jump when the mailto link is finally rendered by
react.

This commit reduces the layout jump by including an empty span tag,
causing the renderer to allocate the approximate space needed.